requirements.txt
Django==3.0.6
PTable==0.9.2
PyNaCl==1.3.0
djangorestframework==3.11.0
pycodestyle==2.6.0
requests~=2.23.0
setuptools==46.4.0
wheel==0.34.2
